+++ src/main/java/com/example/jjangushrine/domain/order/dto/response/OrderItemRes.java
@@ -2,7 +2,7 @@ package com.example.jjangushrine.domain.order.dto.response;
 
 public record OrderItemRes(
         String productName,
-        int quantity,
+        Short quantity,
         int productPrice
 ) {
 }
